Keep classic WinUSB on device-level MS OS and harden the ISP session.
Win11 A/B showed nested MS OS 0xAE left classic unbound while 0x9E binds;
reset prog state on CONNECT/DISCONNECT, tighten wire types and nbytes=0
guards, stop advertising TPI until silicon proof, and record ACCEPTANCE-001.

+++ firmware/include/usbasp/isp.h
@@ -18,11 +18,11 @@ void isp_out_clr_bit(uchar bit);
 uchar ispTransmit_sw(uchar send_byte);
 uchar ispTransmit_hw(uchar send_byte);
 uchar ispEnterProgrammingMode(void);
-uchar ispReadEEPROM(unsigned int address);
-uchar ispWriteFlash(unsigned long address, uchar data, uchar pollmode);
-uchar ispFlushPage(unsigned long address, uchar pollvalue);
-uchar ispReadFlash(unsigned long address);
-uchar ispWriteEEPROM(unsigned int address, uchar data);
+uchar ispReadEEPROM(uint16_t address);
+uchar ispWriteFlash(uint32_t address, uchar data, uchar pollmode);
+uchar ispFlushPage(uint32_t address, uchar pollvalue);
+uchar ispReadFlash(uint32_t address);
+uchar ispWriteEEPROM(uint16_t address, uchar data);
 
 typedef uchar (*isp_transfer_fn)(uchar);
